Se reestructura proyecto para usar protected
